+++ CHANGELOG.md
@@ -89,6 +89,7 @@
 - Enh #90: Add `HasFontWeight` trait and corresponding tests for managing SVG `font-weight` attribute (@terabytesoftw)
 - Enh #91: Add `HasLengthAdjust` trait and corresponding tests for managing SVG `lengthAdjust` attribute (@terabytesoftw)
 - Enh #92: Add `HasLetterSpacing` trait and corresponding tests for managing SVG `letter-spacing` attribute (@terabytesoftw)
+- Enh #93: Add `HasRotate` trait and corresponding tests for managing SVG `rotate` attribute (@terabytesoftw)
 
 ## 0.2.0 March 31, 2024